Crate shipcat_definitions

Re-exports

pub use crate::region::Region;

pub use crate::region::VaultConfig;

pub use crate::region::VersionScheme;

pub use crate::region::KongConfig;

pub use crate::config::Config;

pub use crate::config::Cluster;

pub use crate::config::Team;

pub use crate::config::ManifestDefaults;

pub use crate::manifest::Manifest;

pub use crate::vault::Vault;

Modules

config
Master config with cross-region data

manifest

math
Computational helpers

region
Config with regional data

structs
Structs for the manifest

template
A renderer of tera templates (jinja style)

vault
A Hashicorp Vault HTTP client using reqwest

Structs

Crd
Basic CRD wrapper struct

CrdEvent
CRD Event wrapper

CrdList
Basic CRD List wrapper struct

Error
The Error type.

Enums

ConfigType
Various states a Config can exist in depending on resolution.

CrdEventType
Types of events returned from watch requests

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

gen_all_crds

Type Definitions

Result
Convenient wrapper around std::Result.