Struct shipcat_definitions::config::Config

pub struct Config {
    pub defaults: ManifestDefaults,
    pub clusters: BTreeMap<String, Cluster>,
    pub contextAliases: BTreeMap<String, String>,
    pub locations: BTreeMap<String, Location>,
    pub slack: SlackParameters,
    pub github: GithubParameters,
    pub allowedLabels: Vec<String>,
    pub allowedCustomMetadata: BTreeSet<String>,
    pub versions: BTreeMap<Environment, Version>,
    pub owners: Owners,
    // some fields omitted
}

Main manifest, serializable from shipcat.conf

Fields
defaults: ManifestDefaults

Global defaults for the manifests

clusters: BTreeMap<String, Cluster>

Cluster definitions

contextAliases: BTreeMap<String, String>

Context aliases, e.g. prod-uk-green -> prod-uk

locations: BTreeMap<String, Location>

Location definitions

slack: SlackParameters

Slack parameters

github: GithubParameters

Gihub parameters

allowedLabels: Vec<String>

Allowed labels

allowedCustomMetadata: BTreeSet<String>versions: BTreeMap<Environment, Version>

Shipcat version pins

owners: Owners

Owners of services, squads, tribes

Populated from teams.yml

Methods

impl Config

pub fn verify(&self) -> Result<()>

pub fn verify_version_pin(&self, env: &Environment) -> Result<()>

pub fn get_appropriate_version_pin(&self, env: &Environment) -> Result<Version>

pub fn bail_on_version_older_than(pin: &Version) -> Result<()>

pub fn print(&self) -> Result<()>

Print Config to stdout

pub fn list_regions(&self) -> Vec<String>

Helper for small utils that don't need the full struct

pub fn resolve_cluster(
&self,
ctx: &str,
cluster: Option<String>
) -> Result<(Cluster, Region)>

Work out the cluster from the kube config

Can be done unambiguously when cluster name is specified, Otherwise we will find the first candidate cluster serving this context and bail if there's more than one.

pub fn has_secrets(&self) -> bool

pub fn get_region(&self, ctx: &str) -> Result<Region>

Region retriever interface

This retieves the region after calling resolve_context. Useful for small helper subcommands that do validation later.

pub fn get_regions(&self) -> Vec<Region>

Region exposer (needed in a few special cases, raftcat, crd reconcile)

pub fn find_owning_cluster(&self, region: &Region) -> Option<Cluster>

Find the Cluster struct that owns this Region

impl Config

Filesystem accessors for Config

These must live in here because they use private methods herein.

pub async fn new<'_>(
state: ConfigState,
context: &'_ str
) -> Result<(Config, Region)>

Main constructor for CLI

Pass this a region request via argument or a current context

pub async fn read() -> Result<Config>

Read a config in pwd and leave placeholders

pub fn has_all_regions(&self) -> bool

pub fn get_region_unchecked(&self, region: &str) -> Option<&Region>

Region retriever for global reducers

Assumes you have not filtered a config in main accidentally.