Struct shipcat_definitions::region::Region

pub struct Region {
    pub name: String,
    pub namespace: String,
    pub environment: Environment,
    pub reconciliationMode: ReconciliationMode,
    pub cluster: String,
    pub versioningScheme: VersionScheme,
    pub base_urls: BTreeMap<String, String>,
    pub env: BTreeMap<String, String>,
    pub kong: Option<KongConfig>,
    pub statuscake: Option<StatuscakeConfig>,
    pub ip_whitelist: Vec<String>,
    pub kafka: KafkaConfig,
    pub vault: VaultConfig,
    pub logzio: Option<LogzIoConfig>,
    pub grafana: Option<GrafanaConfig>,
    pub sentry: Option<SentryConfig>,
    pub locations: Vec<String>,
    pub webhooks: Vec<Webhook>,
    pub customResources: Option<CRSettings>,
    pub defaults: DefaultConfig,
    pub destinationRuleHostRegex: Option<Regex>,
}

A region is an abstract kube context

Either it's a pure kubernetes context with a namespace and a cluster, or it's an abstract concept with many associated real kubernetes contexts.

Fields
name: String

Name of region

namespace: String

Kubernetes namespace

environment: Environment

Environment (e.g. dev or staging)

reconciliationMode: ReconciliationMode

Reconciliation mode

This affects how cluster crd reconcile behaves in the region.

cluster: String

Primary cluster serving this region

Shipcat does not use this for to decide where a region gets deployed, but it is used to indicate where the canonical location of a cluster is.

During blue/green cluster failovers the value of this string may not be accurate.

Jobs that decide where to deploy a region to should use get clusterinfo with explicit cluster names and regions.

versioningScheme: VersionScheme

Versioning scheme

base_urls: BTreeMap<String, String>

Important base urls that can be templated in evars

env: BTreeMap<String, String>

Environment variables to inject

kong: Option<KongConfig>

Kong configuration for the region

statuscake: Option<StatuscakeConfig>

Statuscake configuration for the region

ip_whitelist: Vec<String>

List of Whitelisted IPs

kafka: KafkaConfig

Kafka configuration for the region

vault: VaultConfig

Vault configuration for the region

logzio: Option<LogzIoConfig>

Logz.io configuration for the region

grafana: Option<GrafanaConfig>

Grafana details for the region

sentry: Option<SentryConfig>

Sentry URL for the region

locations: Vec<String>

List of locations the region serves

webhooks: Vec<Webhook>

All webhooks

customResources: Option<CRSettings>

CRD tuning

defaults: DefaultConfig

Default values for services

destinationRuleHostRegex: Option<Regex>

The regular expression used to verify destination rules' regions

Methods

impl Region

pub async fn secrets<'_>(&'_ mut self) -> Result<()>

pub async fn verify_secrets_exist<'_>(&'_ self) -> Result<()>

pub fn vault_url(&self, app: &str) -> String

pub fn grafana_url(&self, app: &str) -> Option<String>

pub fn sentry_url(&self, slug: &str) -> Option<String>

pub fn logzio_url(&self, app: &str) -> Option<String>

pub fn raftcat_url(&self) -> Option<String>