Module shipcat_definitions::vault

A Hashicorp Vault HTTP client using reqwest

Structs

ListSecrets
List data retrieved from Vault when listing available secrets

Secret
Secret data retrieved from Vault using only standard fields

Vault
Vault client with cached data

Enums

Mode
Vault usage mode

SecretValue
Secrets in vault values can be integers or strings

Constants

_IMPL_DESERIALIZE_FOR_ListSecrets

_IMPL_DESERIALIZE_FOR_Secret

_IMPL_DESERIALIZE_FOR_SecretValue

Functions

default_addr

default_token